requirements.txt
Django==5.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.0.0
pytest-django==4.8.0
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
django-stubs==4.2.7
coreapi==2.3.3 